+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Upcoming
 * Fix bug that removes "insecure" option from conf file
 * Allow updating findings
+* Clean up NOTICE file
 
 # 0.31
 * Allow usage of custom CAs